requirements.txt
alembic==1.2.1
aniso8601==8.0.0
attrs==19.3.0
Flask==0.10.1
Flask-JWT==0.3.2
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
more-itertools==7.2.0
passlib==1.7.1
PyJWT==1.4.2
PyMySQL==0.9.3
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
SQLAlchemy-Utils==0.34.2
Werkzeug==0.16.0
zipp==0.6.0
google-cloud-storage
pyimgur
boto3
flask-cors